# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /pinata/plot-assets:
    post:
      operationId: uploadPlotAssets
      tags: [pinata]
      summary: Upload DHA plot image and metadata to Pinata IPFS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlotAssetsUploadRequest"
      responses:
        "201":
          description: IPFS assets uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlotAssetsUploadResponse"
        "400":
          description: Invalid upload request
        "502":
          description: Pinata upload failed
        "503":
          description: Pinata is not configured
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PlotAssetsUploadRequest:
      type: object
      properties:
        metadata:
          type: object
          additionalProperties: true
        image:
          type: object
          properties:
            dataUrl:
              type: string
            filename:
              type: string
            contentType:
              type: string
          required:
            - dataUrl
      required:
        - metadata
    PlotAssetsUploadResponse:
      type: object
      properties:
        metadataUri:
          type: string
        imageUri:
          type:
            - string
            - "null"
        metadataCid:
          type: string
      required:
        - metadataUri
        - imageUri
        - metadataCid